Add loading of data for specific UID for operator

+++ scripts/operator/centres.js
@@ -3,19 +3,27 @@ import {
   activateEventForSelectedCentreItem,
 } from "./centres-display.js";
 import { assignEventHandlers } from "./centres-assign-handlers.js";
+import { auth, onAuthStateChanged } from "../firebase/authentication.js";
 
 document.addEventListener("DOMContentLoaded", async () => {
   const urlParams = new URLSearchParams(window.location.search);
   const centreId = urlParams.get("centreId");
 
-  // TODO get UID
-  await renderSidebar("OPERATOR_001");
-  await assignEventHandlers();
-  sessionStorage.clear();
+  onAuthStateChanged(auth, async (user) => {
+    if (user) {
+      const uid = user.uid;
+      await renderSidebar(uid);
+      await assignEventHandlers();
+      sessionStorage.clear();
 
-  if (centreId) {
-    activateEventForSelectedCentreItem(centreId);
-  } else {
-    document.querySelector("ul.sidebar__menu").firstElementChild.click();
-  }
+      if (centreId) {
+        activateEventForSelectedCentreItem(centreId);
+      } else {
+        document.querySelector("ul.sidebar__menu").firstElementChild.click();
+      }
+    } else {
+      alert("You are signed out. Redirecting to HawkPortal login page.");
+      window.location.href = "./signin.html";
+    }
+  });
 });
